LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching-distributed-text-node-expected.txt

Tests to ensure that event dispatching behaves as the Shadow DOM spec describes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




Composed Shadow Tree will be:
DIV	 id=sandbox
	DIV	 id=top
		DIV	 id=shadow-host
			#text	


Moving mouse from a distributed text node to top

  mouseout
     @shadow-host (target: shadow-host) (related: top)
     @top (target: shadow-host) (related: top)

  mouseover
     @top (target: top) (related: shadow-host)

  touchstart
     @content (target: content)
     @shadow-root (target: content)
     @shadow-host (target: shadow-host)
     @top (target: shadow-host)

  mousewheel
     @content (target: content)
     @shadow-root (target: content)
     @shadow-host (target: shadow-host)
     @top (target: shadow-host)
PASS successfullyParsed is true

TEST COMPLETE
Text Node
